#22 created breakpoints in the tailwind configuration
