+++ app/components/friends/page.tsx
@@ -1,10 +1,16 @@
-import Friend from "./Friend";
+import Friend from "../../../components/Friend";
 import "../../globals.css";
 
 const Friends = () => (
   <div>
     <h1 className="normal-case font-bold">Friends</h1>
     <div className="divider" />
+    <Friend />
+    <Friend />
+    <Friend />
+    <Friend />
+    <Friend />
+    <Friend />
     <Friend />
   </div>
 );
