+++ app/components/notifications/page.tsx
@@ -1,10 +1,14 @@
-import Notification from "./Notification";
+import Notification from "../../../components/Notification";
 import "../../globals.css";
 
 const Notifications = () => (
   <div>
     <h1 className="normal-case font-bold">Notifications</h1>
     <div className="divider" />
+    <Notification />
+    <Notification />
+    <Notification />
+    <Notification />
     <Notification />
   </div>
 );
