--- app/friends/page.tsx
@@ -1,12 +1,0 @@
-import Friend from "./Friend";
-import "../globals.css";
-
-const Friends = () => (
-  <div>
-    <h1 className="normal-case font-bold">Friends</h1>
-    <div className="divider-vertical" />
-    <Friend />
-  </div>
-);
-
-export default Friends;
