--- app/notifications/page.tsx
@@ -1,12 +1,0 @@
-import Notification from "./Notification";
-import "../globals.css";
-
-const Notifications = () => (
-  <div>
-    <h1 className="normal-case font-bold">Notifications</h1>
-    <div className="divider-vertical" />
-    <Notification />
-  </div>
-);
-
-export default Notifications;
