+++ components/Friend.tsx
@@ -1,0 +1,16 @@
+import "../app/globals.css";
+
+const Friend = () => (
+    <div>
+    <div className="h-fit">
+      <div className="flex flex-row justify-between items-center">
+        <div><h1 className="normal-case font-bold">Name</h1></div>
+        <div><h1 className="normal-case">1 hr</h1></div>
+      </div>
+      <div>Song - Artist</div>
+    </div>
+    <div className="divider" />
+  </div>
+);
+
+export default Friend;
