+++ components/Notification.tsx
@@ -1,13 +1,13 @@
-import "../../globals.css";
+import "../app/globals.css";
 
 const Notification = () => (
     <div>
     <div className="h-fit">
       <div className="flex flex-row justify-between items-center">
-        <div><h1 className="normal-case font-bold">New Release</h1></div>
-        <div><h1 className="normal-case font-bold">12:00 PM</h1></div>
+        <div><h1 className="normal-case font-bold">Title</h1></div>
+        <div><h1 className="normal-case font-bold">HH:MM PM</h1></div>
       </div>
-      <div>New Release from Tame Impala.</div>
+      <div>Body</div>
     </div>
     <div className="divider" />
   </div>
