+++ package.json
@@ -23,6 +23,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "reactfire": "^4.2.2",
     "tailwindcss": "^3.2.4"
   },
   "devDependencies": {
