+++ tailwind.config.js
@@ -2,6 +2,22 @@
 module.exports = {
   content: ["./app/**/*.tsx", "./components/**/*.tsx"],
   theme: {
+    screens: {
+      'sm': '640px',
+      // => @media (min-width: 640px) { ... }
+
+      'md': '768px',
+      // => @media (min-width: 768px) { ... }
+
+      'lg': '1024px',
+      // => @media (min-width: 1024px) { ... }
+
+      'xl': '1280px',
+      // => @media (min-width: 1280px) { ... }
+
+      '2xl': '1536px',
+      // => @media (min-width: 1536px) { ... }
+    },
     extend: {},
   },
   // eslint-disable-next-line global-require
